api/check_email/requirements.txt
functions-framework==3.*
pandas
google-cloud-storage